requirements.txt
cryptography>=36.0.0
pycryptodome>=3.15.0
psutil>=5.9.0
pywin32>=303;platform_system=="Windows"
pathlib2>=2.3.6
tqdm>=4.64.0
argparse>=1.4.0 